Question 36 answered: parentBoneIndex is one byte, and cf7 hid it
readControlinator read parentBoneIndex with s.i32(). It is one byte. With
COMPRESSED_INTEGERS an i32 of 0 is a single varint byte, so the call was
right by accident on every file this project has ever parsed -- all cf7 --
and ate four bytes in a cf0 chunk from the public archive, leaving the reader
three past the next Thing's 0xAA.

That is the general lesson and it is worth more than the fix: cf7 hides
field-width bugs, any field that is usually zero reads correctly whatever
width it is declared with, and until today there was no cf0 corpus at all. It
is now in sequencer-data-model.md, and the 32 archived chunks are that
corpus.

How it was found is the part to copy, because the 0xAA marker never names the
byte that is wrong. setTrace put the failure inside CONTROLINATOR; wrapping
every Serializer method logged each read with its VALUE, not just its width,
because widths align at any offset; SWITCH's tail read plausibly -- radius
250, 18 outputs, angleRange 180, randomOnTimeMin and Max both 30 -- and the
next field did not: parentBoneIndex = 4,161,536. That is 0x003F8000, a zero
byte followed by three quarters of 3F 80 00 00, which is 1.0f. The raw bytes
settled it: an identity matrix starts at 65641 with its four 1.0s exactly on
m00, m05, m10 and m15, and CONTROLINATOR ends at 65707, which is where the
-1417.5 / +1417.5 min-max pair after it begins.

  archived chunks parsing   31 of 32  ->  32 of 32
  island problems                  1  ->  0
  Things recovered from them       -  ->  5,832
  golden fixture            62,158 placements byte for byte, unchanged

readChunkIslands is now exported and readChunk is written in terms of it, so
the next probe into one island does not start by editing level.ts.

Open questions: 13 -> 12.

+++ src/core/parts.ts
@@ -1257,7 +1257,16 @@ export function readControlinator(s: Serializer, readers: ReadonlyMap<string, Pa
   if (subVersion > 0x6e) s.bool(); // killRiderOnCreatureDeath
   thing(); // padSwitch
   if (subVersion > 0x4b) {
-    s.i32(); // parentBoneIndex
+    // ⚠️ **One byte, not an `i32`, and the difference only shows in an
+    // uncompressed stream.** With `COMPRESSED_INTEGERS` an `i32` of 0 is a
+    // single varint byte, so `s.i32()` read this correctly on the whole corpus
+    // -- every file in it is `cf7`. In a `cf0` chunk out of the public archive
+    // the same call ate four bytes and put the reader three past the next
+    // Thing's marker. Measured on chunk `9712345a…` island 48: the field is the
+    // `00` at 65640 and `parentBoneOffset` begins at 65641 with `3f 80 00 00`,
+    // an identity matrix whose four 1.0s land exactly on m00, m05, m10 and m15.
+    // See question 36 in steering/answered-questions.md.
+    s.u8(); // parentBoneIndex
     s.matrix(); // parentBoneOffset
   }
   if (version > 0x3ec) s.u8(); // playerMode
